Move anonymize functions into a post_processor.
